attempt No. 1 to add new SQL files.

+++ src/main/java/com/revature/models/Wizards.java
@@ -102,7 +102,7 @@ public class Wizards implements Serializable{
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + wizardId;
+		result = (int) (prime * result + wizardId);
 		result = prime * result + ((firstName == null) ? 0 : firstName.hashCode());
 		result = prime * result + ((lastName == null) ? 0 : lastName.hashCode());
 		result = prime * result + ((wandWood == null) ? 0 : wandWood.hashCode());
